+++ packages/esbuild-plugin/src/index.ts
@@ -1,3 +1,5 @@
+import { dirname, join } from 'path';
+
 import {
   cssFileFilter,
   virtualCssFileFilter,
@@ -6,6 +8,7 @@ import {
   compile,
   vanillaExtractFilescopePlugin,
   IdentifierOption,
+  CompileOptions,
 } from '@vanilla-extract/integration';
 import type { Plugin } from 'esbuild';
 
@@ -13,10 +16,14 @@ const vanillaCssNamespace = 'vanilla-extract-css-ns';
 
 interface VanillaExtractPluginOptions {
   outputCss?: boolean;
+  /**
+   * @deprecated Use `esbuildOptions.external` instead.
+   */
   externals?: Array<string>;
   runtime?: boolean;
   processCss?: (css: string) => Promise<string>;
   identifiers?: IdentifierOption;
+  esbuildOptions?: CompileOptions['esbuildOptions'];
 }
 export function vanillaExtractPlugin({
   outputCss,
@@ -24,6 +31,7 @@ export function vanillaExtractPlugin({
   runtime = false,
   processCss,
   identifiers,
+  esbuildOptions,
 }: VanillaExtractPluginOptions = {}): Plugin {
   if (runtime) {
     // If using runtime CSS then just apply fileScopes to code
@@ -43,24 +51,41 @@ export function vanillaExtractPlugin({
       build.onLoad(
         { filter: /.*/, namespace: vanillaCssNamespace },
         async ({ path }) => {
-          let { source } = getSourceFromVirtualCssFile(path);
+          let { source, fileName } = await getSourceFromVirtualCssFile(path);
 
           if (typeof processCss === 'function') {
             source = await processCss(source);
           }
 
+          const rootDir = build.initialOptions.absWorkingDir ?? process.cwd();
+
+          const resolveDir = dirname(join(rootDir, fileName));
+
           return {
             contents: source,
             loader: 'css',
+            resolveDir,
           };
         },
       );
 
       build.onLoad({ filter: cssFileFilter }, async ({ path }) => {
+        const combinedEsbuildOptions = { ...esbuildOptions } ?? {};
+
+        // To avoid a breaking change this combines the `external` option from
+        // esbuildOptions with the pre-existing externals option.
+        if (externals) {
+          if (combinedEsbuildOptions.external) {
+            combinedEsbuildOptions.external.push(...externals);
+          } else {
+            combinedEsbuildOptions.external = externals;
+          }
+        }
+
         const { source, watchFiles } = await compile({
           filePath: path,
-          externals,
           cwd: build.initialOptions.absWorkingDir,
+          esbuildOptions: combinedEsbuildOptions,
         });
 
         const contents = await processVanillaFile({