+++ packages/integration/src/addFileScope.ts
@@ -1,33 +1,31 @@
 import { posix, relative, sep } from 'path';
-import type { PackageInfo } from './packageInfo';
 
 interface AddFileScopeParams {
   source: string;
   filePath: string;
-  packageInfo: PackageInfo;
+  rootPath: string;
+  packageName: string;
 }
 export function addFileScope({
   source,
   filePath,
-  packageInfo,
+  rootPath,
+  packageName,
 }: AddFileScopeParams) {
+  // Encode windows file paths as posix
+  const normalizedPath = posix.join(...relative(rootPath, filePath).split(sep));
+
   if (source.indexOf('@vanilla-extract/css/fileScope') > -1) {
-    return { source, updated: false };
+    return source.replace(
+      /setFileScope\(((\n|.)*?)\)/,
+      `setFileScope("${normalizedPath}", "${packageName}")`,
+    );
   }
 
-  // Encode windows file paths as posix
-  const normalizedPath = posix.join(
-    ...relative(packageInfo.dirname, filePath).split(sep),
-  );
-
-  const packageName = packageInfo.name ? `"${packageInfo.name}"` : 'undefined';
-
-  const contents = `
+  return `
     import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
-    setFileScope("${normalizedPath}", ${packageName});
+    setFileScope("${normalizedPath}", "${packageName}");
     ${source}
     endFileScope();
   `;
-
-  return { source: contents, updated: true };
 }