+++ examples/webpack-react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # vanilla-extract-example-webpack-react
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`8467fc2`](https://github.com/seek-oss/vanilla-extract/commit/8467fc28707372f30d8b6239580244c06859a605)]:
+  - @vanilla-extract/css@1.7.2
+  - @vanilla-extract/babel-plugin@1.1.7
+  - @vanilla-extract/webpack-plugin@2.1.11
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`e531251`](https://github.com/seek-oss/vanilla-extract/commit/e531251689b8795eebd316ae8385f1ecc5b9b8a0)]:
+  - @vanilla-extract/css@1.7.1
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`32f309f`](https://github.com/seek-oss/vanilla-extract/commit/32f309faad90e927efc7a277d4208480c8b5122c), [`69b8460`](https://github.com/seek-oss/vanilla-extract/commit/69b846015ecada3354ea292e1276e5da560923c6)]:
+  - @vanilla-extract/css@1.7.0
+  - @vanilla-extract/webpack-plugin@2.1.8
+
 ## 0.0.7
 
 ### Patch Changes