+++ fixtures/sprinkles/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @fixtures/sprinkles
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`8467fc2`](https://github.com/seek-oss/vanilla-extract/commit/8467fc28707372f30d8b6239580244c06859a605)]:
+  - @vanilla-extract/css@1.7.2
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`e531251`](https://github.com/seek-oss/vanilla-extract/commit/e531251689b8795eebd316ae8385f1ecc5b9b8a0)]:
+  - @vanilla-extract/css@1.7.1
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`98ab94a`](https://github.com/seek-oss/vanilla-extract/commit/98ab94a99fa7dbee395a7ae6ea4af24c7b1bd7ab)]:
+  - @vanilla-extract/sprinkles@1.4.1
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`32f309f`](https://github.com/seek-oss/vanilla-extract/commit/32f309faad90e927efc7a277d4208480c8b5122c)]:
+  - @vanilla-extract/css@1.7.0
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`81a3d98`](https://github.com/seek-oss/vanilla-extract/commit/81a3d98b3fb36660406a59049e1cd464418b00fc)]:
+  - @vanilla-extract/sprinkles@1.4.0
+
 ## 0.0.5
 
 ### Patch Changes