+++ test-helpers/CHANGELOG.md
@@ -1,5 +1,124 @@
 # @vanilla-extract-private/test-helpers
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [[`02ea735`](https://github.com/seek-oss/vanilla-extract/commit/02ea73552d1c1b9b7b9ec44be00d89413e28a97f)]:
+  - @vanilla-extract/vite-plugin@3.2.1
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [[`e373b51`](https://github.com/seek-oss/vanilla-extract/commit/e373b51bfa8401e0746596adafbda350c9fad4c3)]:
+  - @vanilla-extract/esbuild-plugin@2.1.0
+  - @vanilla-extract/vite-plugin@3.2.0
+  - @vanilla-extract/babel-plugin@1.1.7
+  - @vanilla-extract/webpack-plugin@2.1.11
+  - @fixtures/features@0.0.8
+  - @fixtures/low-level@0.0.8
+  - @fixtures/recipes@0.0.9
+  - @fixtures/sprinkles@0.0.10
+  - @fixtures/themed@0.0.8
+  - @fixtures/unused-modules@1.0.8
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [[`9b6f3ea`](https://github.com/seek-oss/vanilla-extract/commit/9b6f3ea4160c84e48172ea01b922888dbc7f7c81), [`162b66f`](https://github.com/seek-oss/vanilla-extract/commit/162b66fb48262182846ff71a25d541e00257672c), [`3fc5040`](https://github.com/seek-oss/vanilla-extract/commit/3fc50406d5e57bb131eaebab42052cb1370cddff)]:
+  - @vanilla-extract/vite-plugin@3.1.7
+
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fixtures/features@0.0.7
+  - @fixtures/low-level@0.0.7
+  - @fixtures/recipes@0.0.8
+  - @fixtures/sprinkles@0.0.9
+  - @fixtures/themed@0.0.7
+  - @fixtures/unused-modules@1.0.7
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fixtures/sprinkles@0.0.8
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [[`f6d5337`](https://github.com/seek-oss/vanilla-extract/commit/f6d5337ae7b955add2bb4a27ffafe1ff55b1a140)]:
+  - @vanilla-extract/vite-plugin@3.1.6
+  - @vanilla-extract/webpack-plugin@2.1.10
+  - @vanilla-extract/esbuild-plugin@2.0.5
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [[`3c9b7d9`](https://github.com/seek-oss/vanilla-extract/commit/3c9b7d9f2f7cba8635e7459c36585109b6616636), [`3c9b7d9`](https://github.com/seek-oss/vanilla-extract/commit/3c9b7d9f2f7cba8635e7459c36585109b6616636)]:
+  - @vanilla-extract/esbuild-plugin@2.0.4
+  - @vanilla-extract/webpack-plugin@2.1.9
+  - @vanilla-extract/vite-plugin@3.1.5
+  - @vanilla-extract/babel-plugin@1.1.6
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`69b8460`](https://github.com/seek-oss/vanilla-extract/commit/69b846015ecada3354ea292e1276e5da560923c6)]:
+  - @vanilla-extract/webpack-plugin@2.1.8
+  - @fixtures/features@0.0.6
+  - @fixtures/low-level@0.0.6
+  - @fixtures/recipes@0.0.7
+  - @fixtures/sprinkles@0.0.7
+  - @fixtures/themed@0.0.6
+  - @fixtures/unused-modules@1.0.6
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`5ebca75`](https://github.com/seek-oss/vanilla-extract/commit/5ebca758d18e7d55d0c079c00178cb14e936ac2e)]:
+  - @vanilla-extract/webpack-plugin@2.1.7
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`6fa65ad`](https://github.com/seek-oss/vanilla-extract/commit/6fa65ad60b5e2aed05f1c3ab01f12c6c8f10a730)]:
+  - @vanilla-extract/vite-plugin@3.1.4
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fixtures/recipes@0.0.6
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`bec1cd8`](https://github.com/seek-oss/vanilla-extract/commit/bec1cd88d78071a995edc76a5c626f361fafcda9)]:
+  - @vanilla-extract/esbuild-plugin@2.0.3
+  - @vanilla-extract/webpack-plugin@2.1.6
+  - @vanilla-extract/babel-plugin@1.1.5
+  - @vanilla-extract/vite-plugin@3.1.3
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fixtures/sprinkles@0.0.6
+
 ## 0.0.7
 
 ### Patch Changes